DAG-1879: Refactor evergreen project configuration extraction
